+++ benchmarks/ensemble_model_overhead/docker-compose.yaml
@@ -1,0 +1,27 @@
+version: "3"
+services:
+  tritonserver:
+    build:
+      context: ./docker
+      dockerfile: Dockerfile
+      target: gpu_image
+    image: getty708/triton-sandbox:24.01-py3-benchmark-ensemble-gpu
+    working_dir: /workspace/triton-sandbox/benchmarks/ensemble_model_overhead
+    command: /bin/bash
+    tty: true
+    volumes:
+      - ../../:/workspace/triton-sandbox
+    environment:
+      PYTHONPATH: /workspace/triton-sandbox
+      TF_ENABLE_ONEDNN_OPTS: 0
+    shm_size: 32g
+    ports:
+      - "8000:8000"
+      - "8001:8001"
+      - "8002:8002"
+    deploy:
+      resources:
+        reservations:
+          devices:
+            - driver: nvidia
+              capabilities: [gpu]
