+++ benchmarks/ensemble_model_overhead/model_repository/ensemble_4/config.pbtxt
@@ -1,0 +1,69 @@
+name: "ensemble_4"
+platform: "ensemble"
+max_batch_size: 0
+input [
+    {
+        name: "image_input"
+        data_type: TYPE_FP32
+        dims: [ -1, 3, 1080, 1920 ]
+    }
+]
+output [
+    {
+        name: "image_output"
+        data_type: TYPE_FP32
+        dims: [ -1, 3, 1080, 1920 ]
+    }
+]
+ensemble_scheduling {
+    step [
+        {
+            model_name: "monolithic_1"
+            model_version: -1
+            input_map {
+                key: "image_input"
+                value: "image_input"
+            }
+            output_map {
+                key: "image_output"
+                value: "image_output_m1"
+            }
+        },
+        {
+            model_name: "monolithic_1"
+            model_version: -1
+            input_map {
+                key: "image_input"
+                value: "image_output_m1"
+            }
+            output_map {
+                key: "image_output"
+                value: "image_output_m2"
+            }
+        },
+        {
+            model_name: "monolithic_1"
+            model_version: -1
+            input_map {
+                key: "image_input"
+                value: "image_output_m2"
+            }
+            output_map {
+                key: "image_output"
+                value: "image_output_m3"
+            }
+        },
+        {
+            model_name: "monolithic_1"
+            model_version: -1
+            input_map {
+                key: "image_input"
+                value: "image_output_m3"
+            }
+            output_map {
+                key: "image_output"
+                value: "image_output"
+            }
+        }
+    ]
+}
