+++ benchmarks/ensemble_model_overhead/utils/const.py
@@ -1,0 +1,4 @@
+from enum import Enum
+
+PIPELINE_ARCH_ENSEMBLE = "ensemble"
+PIPELINE_ARCH_MONOLITHIC = "monolithic"
