--- triton_benchmarks/ensemble_model_v0/docker-compose.yaml
@@ -1,24 +1,0 @@
-version: "3"
-services:
-  tritonserver:
-    image: getty708/triton-sandbox:24.01-py3-benchmark-ensemble-gpu
-    container_name: triton-benchmark-ensemble
-    working_dir: /workspace/triton-sandbox/triton_benchmarks/ensemble_model_v0
-    command: /bin/bash
-    tty: true
-    volumes:
-      - ../../:/workspace/triton-sandbox
-    environment:
-      PYTHONPATH: /workspace/triton-sandbox
-      TF_ENABLE_ONEDNN_OPTS: 0
-    shm_size: 32g
-    ports:
-      - "8000:8000"
-      - "8001:8001"
-      - "8002:8002"
-    deploy:
-      resources:
-        reservations:
-          devices:
-            - driver: nvidia
-              capabilities: [gpu]
