--- triton_benchmarks/ensemble_model_v0/model_repository/dummy_model/config.pbtxt
@@ -1,39 +1,0 @@
-name: "dummy_model"
-backend: "python"
-max_batch_size: 0
-input [
-    {
-        name: "image_input"
-        data_type: TYPE_FP32
-        dims: [ -1, 3, 1080, 1920 ]
-    }
-]
-output [
-    {
-        name: "image_output"
-        data_type: TYPE_FP32
-        dims: [ -1, 3,  1080, 1920 ]
-    }
-]
-instance_group [
-    {
-      count: 1
-      kind: KIND_AUTO
-    }
-]
-model_warmup [
-    {
-        name : "Warmup Sample"
-        batch_size: 1
-        inputs {
-            key: "image_input"
-            value: {
-                data_type: TYPE_FP32
-                dims: [ 8, 3, 1080, 1920 ]
-                random_data: true
-            }
-        }
-        count: 5
-    }
-]
-parameters: { key: "FORCE_CPU_ONLY_INPUT_TENSORS" value: {string_value:"no"}}
