--- triton_benchmarks/ensemble_model_v0/model_repository/pipeline_2/config.pbtxt
@@ -1,45 +1,0 @@
-name: "pipeline_2"
-platform: "ensemble"
-max_batch_size: 0
-input [
-    {
-        name: "image_input"
-        data_type: TYPE_FP32
-        dims: [ -1, 3, 1080, 1920 ]
-    }
-]
-output [
-    {
-        name: "image_output"
-        data_type: TYPE_FP32
-        dims: [ -1, 3, 1080, 1920 ]
-    }
-]
-ensemble_scheduling {
-    step [
-        {
-            model_name: "dummy_model"
-            model_version: -1
-            input_map {
-                key: "image_input"
-                value: "image_input"
-            }
-            output_map {
-                key: "image_output"
-                value: "image_output_m1"
-            }
-        },
-        {
-            model_name: "dummy_model"
-            model_version: -1
-            input_map {
-                key: "image_input"
-                value: "image_output_m1"
-            }
-            output_map {
-                key: "image_output"
-                value: "image_output"
-            }
-        }
-    ]
-}
